+++ previous-versions/common/pkg/models/operations/getintegrationsbranding.go
@@ -24,8 +24,6 @@ type GetIntegrationsBrandingResponse struct {
 	Branding *shared.Branding
 	// HTTP response content type for this operation
 	ContentType string
-	// Your API request was not properly authorized.
-	ErrorMessage *shared.ErrorMessage
 	// HTTP response status code for this operation
 	StatusCode int
 	// Raw HTTP response; suitable for custom response parsing
@@ -44,13 +42,6 @@ func (o *GetIntegrationsBrandingResponse) GetContentType() string {
 		return ""
 	}
 	return o.ContentType
-}
-
-func (o *GetIntegrationsBrandingResponse) GetErrorMessage() *shared.ErrorMessage {
-	if o == nil {
-		return nil
-	}
-	return o.ErrorMessage
 }
 
 func (o *GetIntegrationsBrandingResponse) GetStatusCode() int {